Crate argmin_checkpointing_file

This crate creates checkpoints on disk for an optimization run.

Saves a checkpoint on disk from which an interrupted optimization run can be resumed. For details on the usage please see the documentation of FileCheckpoint or have a look at the example.

Usage

Add the following line to your dependencies list:

[dependencies]
argmin-checkpointing-file = "0.1.0"

License

Licensed under either of

Apache License, Version 2.0, (LICENSE-APACHE or http://www.apache.org/licenses/LICENSE-2.0)

MIT License (LICENSE-MIT or http://opensource.org/licenses/MIT)

at your option.

Contribution

Unless you explicitly state otherwise, any contribution intentionally submitted for inclusion in the work by you, as defined in the Apache-2.0 license, shall be dual licensed as above, without any additional terms or conditions.

Structs

FileCheckpoint

Handles saving a checkpoint to disk as a binary file.

Enums

CheckpointingFrequency

Defines at which intervals a checkpoint is saved.

Traits

Checkpoint

An interface for checkpointing methods